+++ src/mcp/tools/crashlytics/list_top_issues.ts
@@ -2,22 +2,19 @@ import { z } from "zod";
 import { tool } from "../../tool";
 import { mcpError, toContent } from "../../util";
 import { listTopIssues } from "../../../crashlytics/listTopIssues";
+import { APP_ID_FIELD } from "./constants";
 
 export const list_top_issues = tool(
   {
     name: "list_top_issues",
     description: "List the top crashes from crashlytics happening in the application.",
     inputSchema: z.object({
-      app_id: z
-        .string()
-        .optional()
-        .describe(
-          "AppId for which the issues list should be fetched. For an Android application, read the mobilesdk_app_id value specified in the google-services.json file for the current package name. For an iOS Application, read the GOOGLE_APP_ID from GoogleService-Info.plist. If neither is available, use the `firebase_list_apps` tool to find an app_id to pass to this tool.",
-        ),
+      app_id: APP_ID_FIELD,
       issue_count: z
         .number()
         .optional()
-        .describe("Number of issues that needs to be fetched. Defaults to 10 if unspecified."),
+        .describe("Number of issues that needs to be fetched. Defaults to 10 if unspecified.")
+        .default(10),
       issue_type: z
         .enum(["FATAL", "NON-FATAL", "ANR"])
         .optional()
@@ -31,15 +28,14 @@ export const list_top_issues = tool(
     },
     _meta: {
       requiresAuth: true,
-      requiresProject: true,
     },
   },
-  async ({ app_id, issue_type, issue_count }, { projectId }) => {
+  async ({ app_id, issue_type, issue_count }) => {
     if (!app_id) return mcpError(`Must specify 'app_id' parameter.`);
 
     issue_type ??= "FATAL";
     issue_count ??= 10;
 
-    return toContent(await listTopIssues(projectId, app_id, issue_type, issue_count));
+    return toContent(await listTopIssues(app_id, issue_type, issue_count));
   },
 );