+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Added prefix support for multi-instance Cloud Functions extension parameters. (#8911)
-- Fixed a bug when `firebase deploy --only dataconnect` doesn't include GQL in nested folders (#8981)
+- Adds additional Crashlytics tools for debugging/analyzing crashes (#9020)